+++ module_infrastructure/src/main/java/project/redis/ratelimiter/fetchratelimiter/FetchRateLimiter.java
@@ -1,8 +1,8 @@
-package project.redis.ratelimiter;
+package project.redis.ratelimiter.fetchratelimiter;
 
 import org.aspectj.lang.ProceedingJoinPoint;
 
-public interface RateLimiter {
+public interface FetchRateLimiter {
     Object tryApiCall(LimitRequestPerTime limitRequestPerTime, ProceedingJoinPoint joinPoint)
             throws Throwable;
